+++ tests/tests.c
@@ -21,7 +21,7 @@ void registerTest(TestFn func, const char* name)
 
 void runTests()
 {
-    PalBool status = false;
+    PalBool status = PAL_FALSE;
     const char* statusString = nullptr;
     for (int32_t i = 0; i < s_Count; i++) {
         palLog(nullptr, "");